asyncio/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
greenlet==3.0.0rc1
idna==3.4
more-itertools==10.1.0
multidict==6.0.4
requests==2.31.0
SQLAlchemy==1.4.48
urllib3==2.0.4
yarl==1.9.2
python-dotenv==1.0.0